+++ epip/fibonacci/exceptions.py
@@ -1,0 +1,13 @@
+"""Fibonacci domain errors."""
+
+
+class FibonacciError(Exception):
+    pass
+
+
+class InvalidFibonacciInputError(FibonacciError):
+    pass
+
+
+class FibonacciVersionError(FibonacciError):
+    pass